adding devices

+++ backend/app/db/models.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey, JSON
+from sqlalchemy import Column, BigInteger, Integer, String, Float, DateTime, ForeignKey, JSON
 from sqlalchemy.sql import func
 from .base import Base
 
 class User(Base):
     __tablename__ = "users"
     
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     first_name = Column(String)
     last_name = Column(String)
     username = Column(String)
@@ -21,7 +21,7 @@ class Referral(Base):
     __tablename__ = "referrals"
     
     id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey("users.user_id"))
+    user_id = Column(BigInteger, ForeignKey("users.user_id"))
     referrer_id = Column(Integer, ForeignKey("users.user_id"))
     created_at = Column(DateTime(timezone=True), server_default=func.now())
 
@@ -29,7 +29,7 @@ class Payment(Base):
     __tablename__ = "payments"
     
     id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey("users.user_id"))
+    user_id = Column(BigInteger, ForeignKey("users.user_id"))
     amount = Column(Float)
     period = Column(Integer)
     device_type = Column(String)
@@ -42,7 +42,7 @@ class Device(Base):
     __tablename__ = "devices"
     
     id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey("users.user_id"))
+    user_id = Column(BigInteger, ForeignKey("users.user_id"))
     device_name = Column(String)
     vpn_key = Column(String)
     outline_key_id = Column(String, nullable=True)
